Add batch package and initial SQL migrations and seed data
This commit adds the batch package to the codebase, which includes a new file `main.go`. It also introduces two new SQL migration files, `20240919030921_init.sql` and `00000000000000_init.sql`, as well as a seed data file `00000000000000_init.sql`. Additionally, a new file `settings.json` is added to the `.vscode` directory, and the `main.go` file is modified to include a title, version, and description.