Refactor cache creation logic and optimize data storage
This commit refactors the CreateCache function in the cache.go file. It introduces two new types, eventType and id, to represent event types and event IDs respectively. The function now takes an additional parameter, anemosData, which is used to create the cache based on the data retrieved from the API.

The changes include:
- Creating a map to store the data to be saved in Redis, using the event ID as the key
- Creating a map to store the event IDs for each event type for the past week
- Creating a list of event types
- Populating the cache data and the weekly event data based on the anemosData

These changes optimize the storage and retrieval of data in the cache.